Added new form of validation, based on statistics
Can now specify as acceptance criterion a statistical test of the distribution of cell count or density over voxels.
The actual value being compared against is going to be the negative logarithm of the p-value.
So use something like 1 or at most 2 for threshold.
To use, give {"log-p-value": thresh} as "tolerance" of your test.